simplify grammar construction
Make the process of defining a grammar simpler. The user may now declare
terminals and nonterminals up front, which yields IDs that can be used
throughout the grammar. This allows future addition of metadata in one
place. The grammar then consists of a list of rules/productions which
map a nonterminal to a sequence of symbols. Multiple rules with the same
left hand side nonterminal may exist.