impl item set closure computation
Implement computing the closure of an item set. Also create the skeleton
of the Honalee algorithm. There seems to be quite some overlap between
computing follow sets and first sets, since both need to potentially
step through a sequence of symbols in the presence of an epsilon. Surely
there's a cleaner way to do this.

Also, the current implementation will just refuse to work if it
encounters a group or optional symbol, which is pretty bad. We should
really handle these cases properly.