change inc_list into an index set
In the Honalee algorithm, change the `inc_list` from being an actual
queue of item sets to an ordered index set. Rather than adding
incomplete item sets to the `inc_list`, they are now added directly to
the `done_list`, and their index is inserted into the `inc_list` set.
This change has no effect right now, but later on will allow item sets
that have merged in another set to be re-updated.