honor kernel items while merging
Fix the item set `merge()` function to honor the kernel items of the
merged set. Detect if an item to be merged in is a kernel item, and if
it is, insert it after the last kernel item rather than just appending
it. Until now all merged-in items would be demoted to regular items.