add state machine repr
Fill in the `machine` module with a representation of a state machine
that is equivalent to a certain grammar. The `StateMachine` struct has
a `try_from` function which performs the conversion from the high-level
`ItemSets` to the states. This function fails if there are any conflicts
in the grammar. It is expected that further optimization and
legalization passes applied to the `ItemSets` are used to ensure that
no conflicts exist.