add basic state code generation
Add the `generate_parser` function to the `backend` module which allows
for code generation. For the time being generate very simple versions of
the parse functions which simply switch over the token and then call a
corresponding function. Reduce actions are not properly handled yet. We
also need some form of runtime library that contains some of the code
common to all generated parsers.