add reduction functions
Extend the backend and `tribble1` example/test to support reduction
functions. The name of these functions may be registered with the
backend, which causes the generated parser code to call these functions
in order to reduce a rule. Add proper symbol destructuring to the
generated parser, such that the reduction functions can operate on the
actual type of symbols that they expect.