fix follow set issue
Fix an issue in the computation of follow sets used for closing item
sets. In left-recursive rules where the recursion would be followed by
a nonterminal, that nonterminal's first set would not display the
appropriate terminals. The fix is to have the implementation of
`FirstSets::get` operate on `NonterminalId` rather than the incorrect
`RuleId`, then adjusting the follow-up errors.