map parsed desc to grammar
Extend the tool to map a parsed grammar description to the actual
underlying grammar data structure. Also offer the possibility to dump
the AST and the computed item sets to stdout. Generate the associated
state machine, but don't produce any output code. This will help detect
any conflicts in the grammar for now.