add action conflict detection
Add the `glr` module which will hold the GLR/LR(k) analysis facilities.
Add the `find_conflicts` function there that analyses sets of items for
conflicting actions to be taken upon the same symbol. This is the basis
for all further analysis. Also add an example which calls this analysis.