add conflict node reduction resolution
Add a `ConflictArc` struct that contains the entire subspace of a parser
related to a conflict. Add a corresponding `ConflictNode` which is part
of the arc and represents a collection of hypothetical parser states and
its stack. Add the `find_conflict_arc` function which is responsible for
computing the arc. Add the `advance_node_to_shift` function which takes
a `ConflictNode` and advances the parser states to a point where each
would result in a shift. This is a preparatory step for lockstepping
through the parser state space to find points of reconvergence or
divergence.