fix issue in conflict arc analysis
Fix an issue in the conflict arc analysis where multiple levels of
reductions would mislead the code to assume the parser has deadlocked.
This was due to the fact that reductions only considered terminals on
the stack to find the correct item set to jump to. To fix this, make
each conflict node keep track of separate lanes, which keeps a detailed
history of the reductions that occurred in between shifts. Reductions
now make use of this information to determine the state of each lane
after a reduction.