add proper logging
Add the `log` and `stderrlog` crate dependencies to provide some useful
and configurable logging capability. Add the `-v`, `-q`, and `-t`
options to the command line utility to control the logging. Change many
of the prints throughout the project to `trace!` or `debug!`. This will
help keep program output clean in general.