make grammar parser self-hosting
Extend the grammar syntax and parser to accept annotation of the backend
types, patterns, and function names. Describe the grammar's grammar in a
file and generate the `parser_states.rs` file based on that. Also add a
script `update_parser.sh` which updates the parser based on the system
perplex installation.