track shift/reduce history in conflict arcs
Add a separate subgraph to conflict arcs which allows tracking of the
shift/reduce history between the lanes. Until now only the stack was
tracked via the `parents`/`seq` fields. Determining the stack up to a
node would involve concatenating the sequences along the node's parents,
yielding multiple possible stacks depending on the fanin of the parents.
Add the `ancestors/reductions` fields which track the previous lanes
that spanwed a lane, and the reductions that led to the lane's current
stack. Finding the sequence of shift/reduce actions up to a node now
consists of concatenating the node's reductions with its ancestors',
interleaving the shift actions between the nodes appropriately.