impl local ambiguity resolution
Implement local ambiguity resolution. To this end, add the function
`resolve_local_ambiguity` to the `glr` module. It detects which rules
of the same nonterminal in a local ambiguity cause a conflict and
isolates the range of symbols that need to be unified. The next step is
to propose a grammar change, but this is not yet implemented.