fix honalee shift/shift conflicts
Fix an issue in the Honalee algorithm where shift/shift conflicts would
be produced, which should never happen. When spawning the transitions
out of an item set that already had some transitions present, the
algorithm would only consider the items without transitions for the new
item set. This caused partial item sets to be constructed, which have
a different set of kernel item cores and thus could not be merged into
previous item sets.