map group/optional/repeat symbols to grammar
Add support for mapping of extended grammar symbols (groups, optionals,
and repetitions) to actual grammar rules. This makes the `parser` module
even more convoluted, so a refactoring might be in order in the near
future. There is currently no way to actually generate a parser for an
extended grammar, since the implicitly added rules cannot be assigned
any reduction functions, and none are synthesized automatically.