add extended grammar
Add module `ext` with facilities to describe extended grammars. That is,
in contrast to the existing `grammar` module, a grammar that is most
convenient to describe as a human rather than most convenient for the
generator to convert into a parser. It allows for grouped, optional,
repeated, and choice symbols, and carries all the metadata needed to
synthesize AST nodes and reduction functions as needed.