add separate ast synth proto tree
Add separate tree nodes for nonterminals, sequences, and symbols to the
AST synthesis procedure. One such node is generated for each
corresponding element in the grammar. The resulting tree is then used to
find recursive rules, and will in the future help to assign names to
nodes and determine which nodes should be inlined. Nonterminal nodes
that have only one rule will be inlined, for example. In a final step,
the nodes of this proto tree should then be converted into a final AST
description.