add synth name hints to grammar desc
Allow grammar descriptions to annotate the desired synthesized AST node
name for nonterminals. Adjust the synthesis code to honor this hint and
generate corresponding names. Also make sure that for suppressed nodes,
the actual external type is registered in the synthesized AST such that
the correct types are inserted upon code generation.