impl ext grammar lowering, use in tool
Implement `ext::Grammar::lower()`, which lowers an extended grammar to
its basic grammar equivalent. In the process, multiple additional helper
rules and nonterminals are created.

In the command line tool, map the parsed grammar to an extended grammar,
perform AST synthesis, then lower to a basic grammar and backend for
further processing.